feat: move network status utility functions to separate utils file
